:recycle: Move `WSIStreamDataset` to `dataset_abc.py` - Move `WSIStreamDataset` to `dataset_abc.py`
