+++ tests/models/test_semantic_segmentation.py
@@ -19,13 +19,10 @@ import yaml
 from click.testing import CliRunner
 
 from tiatoolbox import cli
-from tiatoolbox.models import SemanticSegmentor
+from tiatoolbox.models import SemanticSegmentor, WSIStreamDataset
 from tiatoolbox.models.architecture import fetch_pretrained_weights
 from tiatoolbox.models.architecture.utils import centre_crop
-from tiatoolbox.models.engine.semantic_segmentor import (
-    IOSegmentorConfig,
-    WSIStreamDataset,
-)
+from tiatoolbox.models.engine.semantic_segmentor import IOSegmentorConfig
 from tiatoolbox.models.models_abc import ModelABC
 from tiatoolbox.utils import env_detection as toolbox_env
 from tiatoolbox.utils import imread, imwrite
