+++ tiatoolbox/models/__init__.py
@@ -8,6 +8,7 @@ from .architecture.mapde import MapDe
 from .architecture.micronet import MicroNet
 from .architecture.nuclick import NuClick
 from .architecture.sccnn import SCCNN
+from .dataset.dataset_abc import WSIStreamDataset
 from .engine.multi_task_segmentor import MultiTaskSegmentor
 from .engine.nucleus_instance_segmentor import NucleusInstanceSegmentor
 from .engine.patch_predictor import (
@@ -20,7 +21,6 @@ from .engine.semantic_segmentor import (
     DeepFeatureExtractor,
     IOSegmentorConfig,
     SemanticSegmentor,
-    WSIStreamDataset,
 )
 
 __all__ = [
